moved Dynamic loader to common jam code
