+++ plugin/javash/build.gradle
@@ -25,7 +25,6 @@ mainClassName = 'com.twosigma.beaker.javash.Main'
 dependencies {
   compile project(':plugin:jvm')
   compile group: 'org.abstractmeta', name: 'compilation-toolbox', version: '0.3.2'
-  compile group: 'org.xeustechnologies', name: 'jcl-core', version: '2.5'
 }
 
 compileJava {
